+++ e2e/transfers/e2e-transfer-suite/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @coinmasters/e2e-transfer-dogecoin
 
+## 1.12.16
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.13.19
+
 ## 1.12.15
 
 ### Patch Changes